requirements.txt
torch
torchvision
pandas
scikit-learn
matplotlib
pillow
